More consistent German
This PR changes some of the German translations so that they are more consistent: Currently some of the sentences are written in a more formal way (using formal speech with Sie-zen), and some are written informally (with Du-zen). This PR changes all the writing to be more informal, as this fits something that personal like a gratitude app better. Also, I removed the (probably automatic?) translation of the app name "Presently" to "Derzeit", because that's not the name of the app (also not in German).